feat(blueprints): move template routes
