+++ app/routes.py
@@ -1569,6 +1569,7 @@ def init_routes(app: Flask) -> None:
     from app.blueprints.status import create_blueprint as create_status_blueprint
     from app.blueprints.stream import create_blueprint as create_stream_blueprint
     from app.blueprints.system import create_blueprint as create_system_blueprint
+    from app.blueprints.templates import create_blueprint as create_templates_blueprint
     from app.blueprints.views import create_blueprint as create_views_blueprint
 
     if not getattr(app, "_network_bp_registered", False):
@@ -1594,6 +1595,10 @@ def init_routes(app: Flask) -> None:
     if not getattr(app, "_media_bp_registered", False):
         app.register_blueprint(create_media_blueprint())
         app._media_bp_registered = True
+
+    if not getattr(app, "_templates_bp_registered", False):
+        app.register_blueprint(create_templates_blueprint())
+        app._templates_bp_registered = True
 
     if not getattr(app, "_assets_bp_registered", False):
         app.register_blueprint(create_assets_blueprint())
@@ -2564,580 +2569,6 @@ def init_routes(app: Flask) -> None:
             return send_conditional_file(os.path.join(path, latest_file), PNG_TTL_SEC)
 
         abort(404)
-
-    @app.route("/upload_nav_icon", methods=["POST"])
-    @login_required
-    def upload_nav_icon():
-        """Upload or choose a navigation logo."""
-
-        choice = request.form.get("logo_choice")
-        if choice in {"img/glimpser_small.png", "img/glimpser.png"}:
-            update_setting("NAV_ICON", choice)
-            flash("Navigation logo updated", "success")
-            return redirect(url_for("settings"))
-
-        if "logo_file" not in request.files:
-            flash("No logo file provided", "error")
-            return redirect(url_for("settings")), 400
-
-        logo_file = request.files["logo_file"]
-        if logo_file.filename == "":
-            flash("No logo file provided", "error")
-            return redirect(url_for("settings")), 400
-
-        if not allowed_filename(
-            logo_file.filename
-        ) or not logo_file.filename.lower().endswith(".png"):
-            flash("Invalid file name", "error")
-            return redirect(url_for("settings")), 400
-
-        with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-            logo_file.save(temp_file.name)
-            try:
-                img = Image.open(temp_file.name)
-                w, h = img.size
-                if h == 0 or not 2 <= w / h <= 10:
-                    os.unlink(temp_file.name)
-                    flash("Invalid aspect ratio", "error")
-                    return redirect(url_for("settings")), 400
-            except Exception:
-                os.unlink(temp_file.name)
-                flash("Invalid image file", "error")
-                return redirect(url_for("settings")), 400
-
-            dest_dir = os.path.join(app.static_folder, "img")
-            os.makedirs(dest_dir, exist_ok=True)
-            dest_name = secure_filename(logo_file.filename)
-            dest_path = os.path.join(dest_dir, dest_name)
-            shutil.move(temp_file.name, dest_path)
-            update_setting("NAV_ICON", f"img/{dest_name}")
-        flash("Navigation logo uploaded", "success")
-        return redirect(url_for("settings"))
-
-    @app.route("/templates", methods=["GET", "POST", "DELETE"])
-    @login_required
-    def manage_templates():
-        if request.method == "POST":
-            data = request.get_json(silent=True) or {}
-            template_name = validate_template_name(str(data.get("name", "")))
-            if template_name is None:
-                abort(404)
-            url = data.get("url", "")
-            if url and ("onvif" in url or urlparse(url).path in {"", "/"}):
-                try:
-                    endpoints = camera_discovery.autodetect_onvif_endpoints(url)
-                    if endpoints.get("snapshot"):
-                        data["url"] = endpoints["snapshot"]
-                    elif endpoints.get("stream"):
-                        data["url"] = endpoints["stream"]
-                except Exception:
-                    pass
-            if template_manager.save_template(template_name, data):
-                return jsonify({"status": "success", "message": "Template saved"})
-
-        elif request.method == "GET":
-            group = request.args.get("group")
-            search_query = request.args.get("search", "").lower()
-            templates = template_manager.get_templates()
-
-            filtered_templates = {}
-            for name, template in templates.items():
-                template_groups = template.get("groups", "").split(",")
-                if (group == "all" or group in template_groups) and (
-                    not search_query
-                    or search_query in name.lower()
-                    or search_query in template.get("url", "").lower()
-                    or any(search_query in g.lower() for g in template_groups)
-                ):
-                    filtered_templates[name] = template
-
-            return jsonify(filtered_templates)
-
-        elif request.method == "DELETE":
-            data = request.get_json(silent=True) or {}
-            template_name = validate_template_name(str(data.get("name", "")))
-            if template_name is None:
-                abort(404)
-            if template_manager.delete_template(template_name):
-                return jsonify({"status": "success", "message": "Template deleted"})
-            else:
-                return (
-                    jsonify({"status": "failure", "message": "Template not found"}),
-                    404,
-                )
-
-    @app.route("/templates/<string:template_name>")
-    @login_required
-    def template_details(template_name: TemplateName):
-        template_name = validate_template_name(str(template_name))
-        if template_name is None:
-            abort(404)
-
-        templates = template_manager.get_templates()
-        template_details = templates.get(template_name)
-        if template_details is None:
-            abort(404)  # Template not found
-        lscreenshots = template_manager.get_screenshots_for_template(template_name)
-        lvideos = template_manager.get_videos_for_template(template_name)
-        object_tokens = ["person", "car", "dog", "cat", "truck", "bus", "bicycle"]
-        return render_template(
-            "template_details.html",
-            template_name=template_name,
-            template_details=template_details,
-            screenshots=lscreenshots,
-            videos=lvideos,
-            object_tokens=object_tokens,
-            clip_model=CLIP_MODEL_NAME,
-            clip_gpu=clip_gpu_available(),
-            page_title="Camera Details",
-        )
-
-    @app.route("/generate_prompt/<string:template_name>", methods=["POST"])
-    @login_required
-    def generate_prompt_route(template_name: TemplateName):
-        """Return a suggested caption prompt for ``template_name``."""
-
-        template_name = validate_template_name(str(template_name))
-        if template_name is None:
-            abort(404)
-
-        prompt = prompt_optimizer.generate_prompt(template_name)
-        return jsonify({"prompt": prompt})
-
-    @app.route("/suggest_fix/<string:template_name>", methods=["POST"])
-    @login_required
-    def suggest_fix_route(template_name: TemplateName):
-        """Return diagnostic info and replacement URL suggestions."""
-
-        template_name = validate_template_name(str(template_name))
-        if template_name is None:
-            abort(404)
-
-        details = template_manager.get_template(template_name)
-        if not details:
-            abort(404)
-
-        url = details.get("url", "")
-        xpaths = [
-            details.get("popup_xpath", ""),
-            details.get("dedicated_xpath", ""),
-        ]
-        info = camera_fix.check_camera_template(url, xpaths)
-        return jsonify(info)
-
-    @app.route("/camera_diagnostics/<string:template_name>")
-    @login_required
-    def camera_diagnostics(template_name: TemplateName) -> Response:
-        """Return live diagnostics for ``template_name``.
-
-        The diagnostics include ping latency, open ports, HTTP banner
-        information and a simple device type label when available.
-        """
-
-        template_name = validate_template_name(str(template_name))
-        if template_name is None:
-            abort(404)
-
-        details = template_manager.get_template(template_name)
-        if not details:
-            abort(404)
-
-        url = details.get("url", "")
-        host = urlparse(url).hostname or url
-        try:
-            ip = socket.gethostbyname(host)
-        except Exception:
-            ip = host
-
-        data: dict[str, typing.Any] = {"ip": ip}
-
-        latency = camera_discovery._ping_latency(ip)
-        if latency is not None:
-            data["ping_ms"] = latency
-
-        ports = camera_discovery._detect_open_ports(ip, camera_discovery.COMMON_PORTS)
-        if ports:
-            data["open_ports"] = ports
-            for p in ports:
-                if p in (80, 8080, 443):
-                    banner = camera_discovery._fetch_http_banner(ip, p)
-                    for k, v in banner.items():
-                        data.setdefault(k, v)
-
-        dtype = camera_discovery._classify_device(
-            {
-                "ip": ip,
-                "protocol": details.get("protocol", urlparse(url).scheme or "http"),
-                "port": details.get("port", urlparse(url).port or 0),
-                "info": data,
-            }
-        )
-        if dtype:
-            data["device_type"] = dtype
-
-        return jsonify(data)
-
-    def delete_setting(name: str) -> bool:
-        name = name.replace("'", "")[:32]
-        if not re.findall(r"^[A-Z_]+?$", name):
-            return False
-
-        session = SessionLocal()
-        try:
-            session.execute(
-                text("DELETE FROM settings WHERE name = :name"), {"name": name}
-            )
-            session.commit()
-        finally:
-            session.close()
-
-        return True
-
-    @app.route("/settings", methods=["GET", "POST"])
-    @login_required
-    def settings():
-        """Render settings page and handle configuration updates via POST."""
-        if request.method == "POST":
-            email_settings = [
-                "EMAIL_ENABLED",
-                "EMAIL_SENDER",
-                "EMAIL_RECIPIENTS",
-                "EMAIL_SMTP_SERVER",
-                "EMAIL_SMTP_PORT",
-                "EMAIL_USE_TLS",
-                "EMAIL_USERNAME",
-                "EMAIL_PASSWORD",
-            ]
-            notification_settings = ["SMS_ENABLED", "CAP_ENABLED"]
-            action = request.form.get("action")
-            if action == "add":
-                new_name = (request.form.get("new_name") or "").strip()
-                new_value = (request.form.get("new_value") or "").strip()
-                if not new_name or not new_value:
-                    flash("Setting name and value are required", "error")
-                    return redirect(url_for("settings")), 400
-                if not re.fullmatch(r"[A-Z_]+", new_name):
-                    flash(
-                        "Setting names must contain only uppercase letters and underscores",
-                        "error",
-                    )
-                    return redirect(url_for("settings")), 400
-                sanitized = validate_setting(new_name, new_value)
-                if sanitized is None:
-                    flash(f"Invalid value for {new_name}", "error")
-                    return redirect(url_for("settings")), 400
-                update_setting(new_name, sanitized)
-            elif action == "delete":
-                name_to_delete = request.form.get("name_to_delete")
-                if name_to_delete:
-                    delete_setting(name_to_delete)
-            elif action == "update_email":
-                for setting in email_settings:
-                    value = request.form.get(setting)
-                    if value is not None:
-                        sanitized = validate_setting(setting, value)
-                        if sanitized is None:
-                            flash(f"Invalid value for {setting}", "error")
-                            return redirect(url_for("settings")), 400
-                        update_setting(setting, sanitized)
-            elif action == "backup":
-                if backup_config():
-                    flash("Configuration backed up successfully", "success")
-                else:
-                    flash("Failed to backup configuration", "error")
-            elif action == "download":
-
-                def generate() -> typing.Generator[bytes, None, None]:
-                    """Stream the configuration JSON after triggering a backup."""
-                    yield b""
-                    backup_config()
-                    if os.path.exists(BACKUP_PATH):
-                        with open(BACKUP_PATH, "rb") as f:
-                            for chunk in iter(lambda: f.read(8192), b""):
-                                yield chunk
-
-                headers = {
-                    "Content-Disposition": "attachment; filename=config_backup.json"
-                }
-                return Response(
-                    stream_with_context(generate()),
-                    mimetype="application/json",
-                    headers=headers,
-                )
-            elif action == "upload":
-                if "file" not in request.files:
-                    flash("No file part", "error")
-                else:
-                    file = request.files["file"]
-                    if file.filename == "":
-                        flash("No selected file", "error")
-                    elif file and allowed_file(file.filename):
-                        file.stream.seek(0, os.SEEK_END)
-                        size = file.stream.tell()
-                        file.stream.seek(0)
-                        if size > MAX_UPLOAD_SIZE:
-                            flash("File exceeds 5 MB limit", "error")
-                        else:
-                            file.save(BACKUP_PATH)
-                            restore_config()
-                            flash(
-                                "Configuration restored successfully",
-                                "success",
-                            )
-                    else:
-                        flash("Invalid file type", "error")
-            elif action == "test_email":
-                send_email_alert(
-                    "Glimpser Test Email",
-                    "This is a test email from Glimpser.",
-                )
-                flash("Email test triggered. Check logs for results.", "info")
-            elif action == "test_sms":
-                send_sms_alert("Test SMS from Glimpser")
-                flash("SMS test triggered. Check logs for results.", "info")
-            elif action == "update_shortcut":
-                path_val = request.form.get("shortcut_path")
-                path = Path(path_val) if path_val else None
-                paths, msg = update_chrome_shortcuts_info(path)
-                if paths:
-                    joined = ", ".join(str(p) for p in paths)
-                    flash(
-                        f"Updated {len(paths)} shortcut{'s' if len(paths) != 1 else ''}: {joined}",
-                        "success",
-                    )
-                else:
-                    flash(f"Failed to update shortcuts: {msg}", "error")
-            else:
-                current = {s["name"]: s["value"] for s in get_all_settings()}
-                bool_settings = {
-                    n for n, v in current.items() if validators.is_bool_string(v)
-                }
-                for name in bool_settings:
-                    if name in email_settings:
-                        continue
-                    new_val = (
-                        "True"
-                        if str(request.form.get(name, "")).lower()
-                        in {"true", "on", "1", "t", "y", "yes"}
-                        else "False"
-                    )
-                    update_setting(name, new_val)
-
-                for name, value in request.form.items():
-                    if (
-                        name
-                        in [
-                            "action",
-                            "new_name",
-                            "new_value",
-                            "name_to_delete",
-                        ]
-                        or name in email_settings
-                        or name in bool_settings
-                    ):
-                        continue
-
-                    sanitized = validate_setting(name, value)
-                    if sanitized is None:
-                        flash(f"Invalid value for {name}", "error")
-                        continue
-
-                    if name in SETTINGS_CHOICES:
-                        update_setting(name, sanitized)
-                        continue
-
-                    update_setting(name, sanitized)
-            flash("Settings updated successfully", "success")
-            return redirect(url_for("settings"))
-
-        settings = get_all_settings()
-        grouped_settings: Dict[str, List[Dict[str, Any]]] = {
-            group: [] for group in SETTINGS_GROUPS
-        }
-        grouped_settings["Other"] = []
-        for setting in settings:
-            placed = False
-            for group, names in SETTINGS_GROUPS.items():
-                if setting["name"] in names:
-                    grouped_settings[group].append(setting)
-                    placed = True
-                    break
-            if not placed:
-                grouped_settings["Other"].append(setting)
-
-        if "Advanced" in grouped_settings:
-            grouped_settings["Advanced"].extend(grouped_settings.get("Other", []))
-        else:
-            grouped_settings["Advanced"] = grouped_settings.get("Other", [])
-        grouped_settings.pop("Other", None)
-
-        # Remove empty groups to avoid blank headings in the UI
-        grouped_settings = {g: items for g, items in grouped_settings.items() if items}
-
-        collapsed_groups = {
-            "Capture",
-            "Admin",
-            "Advanced",
-        }
-        file_location_items = [s for s in settings if s["name"] in FILE_LOCATION_NAMES]
-        file_info = file_location_metrics(file_location_items)
-
-        metrics = scheduling.get_system_metrics()
-        feeds = scheduling.get_feed_status()
-        last_summary = scheduling.get_last_summary_time()
-        log_summary = scheduling.get_or_generate_log_summary()
-        danger_enabled = config.get_setting("DANGER_MODE", "True") == "True"
-        cost_summary, total_tokens, total_cost, total_calls = (
-            template_manager.get_llm_cost_summary()
-        )
-        cost_summary = template_manager.group_cost_summary(cost_summary, top=10)
-
-        chrome_path = get_chrome_path()
-        shortcut_opts = [
-            str(p) for p in LINUX_PATHS if p.exists() and os.access(p, os.W_OK)
-        ]
-        danger_info = {
-            "browser": os.path.basename(chrome_path) if chrome_path else "N/A",
-            "path": chrome_path or "N/A",
-            "version": (get_chrome_version(chrome_path) if chrome_path else "N/A"),
-            "shortcut": str(first_shortcut_path() or "N/A"),
-            "patched": not shortcuts_need_patch(),
-            "running": is_chrome_debug_port_open("127.0.0.1", config.DANGER_PORT),
-            "port": config.DANGER_PORT,
-        }
-
-        last_backup = None
-        if os.path.exists(BACKUP_PATH):
-            ts = datetime.fromtimestamp(os.path.getmtime(BACKUP_PATH))
-            last_backup = ts.strftime("%Y-%m-%d %H:%M:%S")
-
-        templates = template_manager.get_templates()
-        existing_urls = {t.get("url"): n for n, t in templates.items() if t.get("url")}
-
-        tooltips = dict(SETTINGS_TOOLTIPS)
-        if not metrics.get("ffmpeg_gpu_support"):
-            tooltips["FFMPEG_HWACCEL"] = "Hardware acceleration not available"
-
-        return render_template(
-            "settings.html",
-            grouped_settings=grouped_settings,
-            collapsed_groups=collapsed_groups,
-            tooltips=tooltips,
-            metrics=metrics,
-            feeds=feeds,
-            last_summary=last_summary,
-            log_summary=log_summary,
-            cost_summary=cost_summary,
-            total_tokens=total_tokens,
-            total_cost=total_cost,
-            total_calls=total_calls,
-            danger_info=danger_info,
-            shortcut_options=shortcut_opts,
-            choices=SETTINGS_CHOICES,
-            boolean_fields=validators.BOOLEAN_SETTINGS,
-            danger_enabled=danger_enabled,
-            numeric_fields=NUMERIC_FIELDS,
-            email_fields=EMAIL_FIELDS,
-            locked_settings=LOCKED_SETTINGS,
-            file_info=file_info,
-            existing_urls=existing_urls,
-            placeholders=SETTINGS_PLACEHOLDERS,
-            last_backup=last_backup,
-            page_title="Settings",
-        )
-
-    def allowed_file(filename):
-        return "." in filename and filename.rsplit(".", 1)[1].lower() == "json"
-
-    @app.route("/update_template/<string:template_name>", methods=["POST"])
-    @login_required
-    def update_template(template_name: TemplateName):
-        template_name = validate_template_name(str(template_name))
-        if template_name is None:
-            abort(404)
-
-        if True:
-            # Extract form data
-
-            updated_data = {
-                "url": request.form.get("url"),
-                "frequency": request.form.get("frequency"),
-                "timeout": request.form.get("timeout"),
-                "notes": request.form.get("notes"),
-                "popup_xpath": request.form.get("popup_xpath"),
-                "dedicated_xpath": request.form.get("dedicated_xpath"),
-                "callback_url": request.form.get("callback_url"),
-                "proxy": request.form.get("proxy"),
-                "auth_username": request.form.get("auth_username"),
-                "auth_password": request.form.get("auth_password"),
-                "rollback_frames": request.form.get("rollback_frames"),
-                "groups": request.form.get("groups"),
-                "object_filter": request.form.get("object_filter"),
-                "object_confidence": request.form.get("object_confidence", 0.5),
-                "motion": request.form.get("motion", 0.2),
-                "invert": request.form.get("invert", "false").lower()
-                in ["true", "1", "t", "y", "yes", "on"],
-                "dark": request.form.get("dark", "false").lower()
-                in ["true", "1", "t", "y", "yes", "on"],
-                "headless": request.form.get("headless", "false").lower()
-                in ["true", "1", "t", "y", "yes", "on"],
-                "stealth": request.form.get("stealth", "false").lower()
-                in ["true", "1", "t", "y", "yes", "on"],
-                "browser": request.form.get("browser", "false").lower()
-                in ["true", "1", "t", "y", "yes", "on"],
-                "livecaption": request.form.get("livecaption", "false").lower()
-                in ["true", "1", "t", "y", "yes", "on"],
-                "danger": request.form.get("danger", "false").lower()
-                in ["true", "1", "t", "y", "yes", "on"],
-            }
-
-            lremoves = []
-            for lkey in updated_data:
-                if updated_data.get(lkey) is None:
-                    lremoves.append(lkey)
-            for lkey in lremoves:
-                del updated_data[lkey]
-
-            # Validate and normalize the incoming form data. Any missing or
-            # out-of-range values are replaced with sensible defaults. A
-            # ``ValueError`` is raised when required fields are absent.
-            try:
-                updated_data = validate_update_data(updated_data)
-            except ValueError as exc:
-                if request.is_json:
-                    return jsonify({"error": str(exc)}), 400
-                flash(str(exc), "error")
-                return redirect("/templates/" + template_name)
-
-            # Update the template in your storage (e.g., JSON file, database)
-            # This assumes you have a function to update templates
-            template_manager.save_template(template_name, updated_data)
-
-            # Stop the old scheduled job and generate a blank frame so the
-            # updated template is clearly segmented in any compiled video.
-            try:
-                scheduling.scheduler.remove_job(template_name)
-            except Exception:
-                pass
-            screenshots.create_blank_frame(template_name)
-            template_manager.get_template(template_name)
-            try:
-                seconds = int(updated_data.get("frequency", 30 * 60))
-                scheduling.scheduler.add_job(
-                    func=scheduling.update_camera,
-                    trigger="interval",
-                    seconds=seconds,
-                    args=[template_name, updated_data],
-                    id=template_name,
-                    replace_existing=True,
-                )
-            except Exception as e:
-                logging.error("job schedule error: %s", e)
-
-            if request.is_json:
-                return jsonify({"message": "Template updated successfully!"})
-
-            return redirect("/templates/" + template_name)
 
     @app.route("/status")
     @login_required
